# A prototype of much faster langfuse ingestion